refactor(svelte): move inspection/ and interval/ into feature directories
Rename-only slice: place inspection and interval modules under
feature directories without merging frame/apply/teardown or
interval/consumption. Update all relative import specifiers and
mirror the test tree. Tooltip remains a public export via a deepened
path; BoundsEditor and bounds-editor keep their basenames.
