+++ packages/svelte/src/lib/inspection/teardown.ts
@@ -4,9 +4,9 @@
  * Callers own reducer dispatch, field clears, coordinator, and focus.
  */
 
-import type { InteractionSource } from "./interaction/interaction.js";
-import { clearInspectionFingerprint } from "./inspection-resolver.js";
-import type { InspectionHostState } from "./plot-surface-inspection-frame.js";
+import type { InteractionSource } from "../interaction/interaction.js";
+import { clearInspectionFingerprint } from "./resolver.js";
+import type { InspectionHostState } from "./frame.js";
 
 // ---- surface blur / outside pointer dismiss ----
