+++ packages/svelte/src/lib/interval/BoundsEditor.svelte
