+++ packages/svelte/src/lib/interval/consumption.ts
@@ -4,7 +4,7 @@ import type {
   PlotInteractionInterval,
   ReadonlyIntervalDomains,
   SemanticIntervalAxis,
-} from "./interaction/interaction.js";
+} from "../interaction/interaction.js";
 
 /** The stable identity surface needed to decide whether stored state is live. */
 interface IntervalConsumptionPanel {
