+++ packages/svelte/src/lib/interval/precise-bounds.ts
@@ -7,7 +7,7 @@ import type {
   BoundsEditorInput,
   BoundsScale,
 } from "./bounds-editor.js";
-import type { SemanticIntervalAxis } from "./interaction/interaction.js";
+import type { SemanticIntervalAxis } from "../interaction/interaction.js";
 
 export interface BoundsEditorInputForScaleOptions {
   readonly axis: BoundsAxis;
