+++ packages/svelte/tests/fixtures/LargePinnedTooltip.svelte
@@ -3,7 +3,7 @@
     PlotDatum,
     PlotInspectionChange,
   } from "../../src/lib/interaction/interaction.js";
-  import Tooltip from "../../src/lib/Tooltip.svelte";
+  import Tooltip from "../../src/lib/inspection/Tooltip.svelte";
 
   const members = Array.from({ length: 1_000 }, (_, index) => ({
     key: `row-${index}`,
